Location CRUD operations added

+++ domain/services/city_service.go
@@ -3,10 +3,12 @@ package services
 import (
 	"City-Pulse-API/domain/entities"
 	"City-Pulse-API/domain/repositories"
+	"errors"
 )
 
 type CityService struct {
-	Repo repositories.CityRepository
+	Repo         repositories.CityRepository
+	LocationRepo repositories.LocationRepository
 }
 
 func (service *CityService) AllCities() ([]entities.City, error) {
@@ -34,6 +36,14 @@ func (service *CityService) CreateCity(city entities.City) (entities.City, error
 }
 
 func (service *CityService) DeleteCity(id string) (entities.City, error) {
+	locations, err := service.LocationRepo.LocationIDsForCity(id)
+	if err != nil {
+		return entities.City{}, err
+	}
+
+	if len(locations) > 0 {
+		return entities.City{}, errors.New("cannot delete city with associated locations")
+	}
 	city, err := service.Repo.DeleteCity(id)
 	if err != nil {
 		return entities.City{}, err
